services/permits/requirements.txt
uvicorn==0.24.0.post1
haystack-ai==2.2.3
elasticsearch-haystack==0.6.0
hayhooks==0.0.15
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==2.1.1
Authlib~=1.3.0
requests~=2.32.3
python-dotenv==1.0.0
boto3==1.34.139
fastapi==0.108.0
pdf2image==1.17.0
Pillow==10.3.0
pdfminer.six==20231228
pyyaml==6.0.1
ocrmypdf==16.4.1
pypdf==4.2.0
json-repair==0.25.3
natsort==8.4.0
pytest==8.0.0
fuzzywuzzy==0.18.0
oauthlib==3.2.2
requests_oauthlib==2.0.0
diff-match-patch==20230430
celery[redis,elasticsearch]==5.4.0
openpyxl==3.1.5
redis==4.6.0
watchdog==4.0.1
pytest-cov==5.0.0
pytest-asyncio==0.23.8
pytest-mock==3.14.0
azure-ai-formrecognizer==3.3.3
azure-core==1.30.2
